Refactor /provider to /providers
