+++ stories/docs/example.mdx
@@ -1,7 +1,7 @@
 import Viewer from "@samvera/clover-iiif/viewer";
 import { Meta } from "@storybook/addon-docs/blocks";
 import { PluginControl, PluginPanel } from "../../src/plugin/index";
-import { UserTokenProvider } from "../../src/provider/index";
+import { UserTokenProvider } from "../../src/providers/index";
 
 <Meta title="Example" />
