+++ stories/docs/welcome.mdx
@@ -23,7 +23,7 @@ Add the plugin to the `Viewer` component:
 ```tsx
 import Viewer from "@samvera/clover-iiif/viewer";
 import { PluginControl, PluginPanel } from "clover-ai";
-import { UserTokenProvider } from "clover-ai/provider";
+import { UserTokenProvider } from "clover-ai/providers";
 
 <Viewer
   iiifContent={
